feat: add modules subtask
